# Clean up and improve CGLxxDevice code