--- a/src/graphics/opengl/gl14device.h
+++ b/src/graphics/opengl/gl14device.h
@@ -139,12 +139,31 @@ public:
     virtual void DrawPrimitives(PrimitiveType type, const VertexCol *vertices,
         int first[], int count[], int drawCount) override;
 
-    unsigned int CreateStaticBuffer(PrimitiveType primitiveType, const Vertex* vertices, int vertexCount) override;
-    unsigned int CreateStaticBuffer(PrimitiveType primitiveType, const VertexTex2* vertices, int vertexCount) override;
-    unsigned int CreateStaticBuffer(PrimitiveType primitiveType, const VertexCol* vertices, int vertexCount) override;
-    void UpdateStaticBuffer(unsigned int bufferId, PrimitiveType primitiveType, const Vertex* vertices, int vertexCount) override;
-    void UpdateStaticBuffer(unsigned int bufferId, PrimitiveType primitiveType, const VertexTex2* vertices, int vertexCount) override;
-    void UpdateStaticBuffer(unsigned int bufferId, PrimitiveType primitiveType, const VertexCol* vertices, int vertexCount) override;
+    unsigned int CreateStaticBuffer(PrimitiveType primitiveType, const Vertex* vertices, int vertexCount) override
+    {
+        return CreateStaticBufferImpl(primitiveType, vertices, vertexCount);
+    }
+    unsigned int CreateStaticBuffer(PrimitiveType primitiveType, const VertexTex2* vertices, int vertexCount) override
+    {
+        return CreateStaticBufferImpl(primitiveType, vertices, vertexCount);
+    }
+    unsigned int CreateStaticBuffer(PrimitiveType primitiveType, const VertexCol* vertices, int vertexCount) override
+    {
+        return CreateStaticBufferImpl(primitiveType, vertices, vertexCount);
+    }
+    void UpdateStaticBuffer(unsigned int bufferId, PrimitiveType primitiveType, const Vertex* vertices, int vertexCount) override
+    {
+        UpdateStaticBufferImpl(bufferId, primitiveType, vertices, vertexCount);
+    }
+    void UpdateStaticBuffer(unsigned int bufferId, PrimitiveType primitiveType, const VertexTex2* vertices, int vertexCount) override
+    {
+        UpdateStaticBufferImpl(bufferId, primitiveType, vertices, vertexCount);
+    }
+    void UpdateStaticBuffer(unsigned int bufferId, PrimitiveType primitiveType, const VertexCol* vertices, int vertexCount) override
+    {
+        UpdateStaticBufferImpl(bufferId, primitiveType, vertices, vertexCount);
+    }
+
     void DrawStaticBuffer(unsigned int bufferId) override;
     void DestroyStaticBuffer(unsigned int bufferId) override;
 
@@ -214,6 +233,11 @@ private:
     //! Disables shadows
     void DisableShadows();
 
+    template <typename Vertex>
+    unsigned int CreateStaticBufferImpl(PrimitiveType primitiveType, const Vertex* vertices, int vertexCount);
+    template <typename Vertex>
+    void UpdateStaticBufferImpl(unsigned int bufferId, PrimitiveType primitiveType, const Vertex* vertices, int vertexCount);
+
 private:
     //! Current config
     DeviceConfig m_config;
@@ -259,14 +283,6 @@ private:
 
     //! Map of framebuffers
     std::map<std::string, std::unique_ptr<CFramebuffer>> m_framebuffers;
-
-    //! Type of vertex structure
-    enum VertexType
-    {
-        VERTEX_TYPE_NORMAL,
-        VERTEX_TYPE_TEX2,
-        VERTEX_TYPE_COL,
-    };
 
     //! Info about static VBO buffers
     struct VboObjectInfo