This pull request is a subset of the commits from #1155. It's mostly code removal and deduplication.

This removes support for GL1.4 display lists as discussed in #1153.

The last commit does less redundant GL state setting, which has a positive effect on performance:

Renderer | Framerate before | Framerate after | Frame time before | Frame time after | Improvement
--- | --- | --- | --- | --- | ---
`gl14` | 712fps | 763fps | 1.40ms | 1.31ms | +7.16%
`gl21` | 735fps | 748fps | 1.36ms | 1.34ms | +1.79%